backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "certifi>=2024.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "livekit>=0.17.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "chromadb>=0.4.22",
    "pypdf>=3.17.0",
]
